agent-core/requirements.txt
anthropic>=0.40.0
slack-bolt>=1.21.0
slack-sdk>=3.33.0
supabase>=2.9.0
requests>=2.32.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
